Prevent missing EOF AST flushes when in spacing context
This manifested itself in the trailing @import "google.com" statement at
the end of the test CSS file not showing up in the AST. That statement is
*specifically* there because it is technically valid to have a single
statement at-rule at the end of the CSS file without a trailing semicolon.

e.g. 'a{rule:value}@import "google.com"' is totally fine