Handle mixed contexts in nested statements more gracefully
Historically I have assumed that within a nested statement block, there
cannot be any heterogeneity of statements (e.g. either there are only
nested at rules or selector blocks). This is wrong and this commit fixes
this.

Furthermore, this bug was caused by some awkward context parent checking
that is very error prone (as evidenced by this bug). This commit also
removes the need for checking the parent context for the double popping
of a context.